windows/requirements.txt
pystray
pillow
torch
